feat(mm): image-base ASLR to 30 bits — place PIE images high, widen the loader
Move the PIE image window from [4 MiB, 8 MiB) to a 4 TiB span above 16 GiB
and randomise the base across 2^30 page-aligned positions: 30 bits, up
from 8.91. That figure was never a policy choice — ASLR_MAX_LOAD_RANDOM_PAGES
was `512 - USER_ASPACE_PREMAP_PAGES`, the 4 KiB slots left in one 2 MiB PD
entry once the premap was subtracted, so log2(480) = 8.91 was a
restatement of the page-table layout. The multi-level page-table walk
removed the single-PD constraint; this spends that freedom.

The image moves to its own high base rather than growing in place: the
user stack (~8 MiB) and heap (16 MiB) are fixed low, and a wide window
anchored at 4 MiB would have swallowed them. Only ELFCLASS64 images take
the high window — an i386 image's addresses and R_386_RELATIVE slots are
32-bit and cannot represent a base above 4 GiB, so those (only the loader
test fixture ships as one) and flat images keep the fixed low base.

The window move surfaced a cluster of latent 32-bit truncations that were
invisible while every user address fit in 32 bits. The signature widening
in the previous pass caught the function boundaries; these were loop
locals, struct fields and return casts the compiler had no reason to
flag. Each one truncated a now-high address to its low 32 bits:

  loader:  dest_va / va_end (segment copy landed at a low address, so the
           ELF load silently fell back to the flat path and ran the raw
           header as code); max_va_end -> a truncated image_end; the
           *out_entry = (uint32_t) cast (the entry point, so a task jumped
           to the low 32 bits of its real _start); the img_end ~0xFFFu
           mask; the per-segment W^X protect loop (so W^X was never applied
           to a high image); the i386 relocator target
  idt:     the demand-paging gate was a fixed [4 MiB, 80 MiB) window, so a
           fault in a high image was never paged and the task was torn
           down silently (a ring-3 fault prints nothing) — replaced with
           the per-task rust_validate_page_fault, which already knows each
           task's real image/heap/stack and cannot over-admit
  syscall:  arm_image_from_user's ubuf; h_sigaltstack's altstack pointer
           (the handler then ran on an unmapped low stack and faulted
           silently); the ipc reply-buffer pointer
  userspace: the sys_signal stub took a uint32_t handler, and three
           callers cast &handler to (unsigned) — the handler registered at
           its low 32 bits and failed the per-task bounds check

Each was found by driving the boot to the fault and reading CR2/RIP, not
by inspection. The recurring shape is a user pointer or address passing
through a uint32_t; the fix is the same everywhere, to carry it at
addr_t/uint64_t.

The ASLR self-test now uses the 64-bit fixture (only ELFCLASS64 goes high)
and asserts a real entropy floor: 8 spawns must land at 8 distinct bases
in the high window spanning more than 1 GiB. A regression that shrank the
window back toward the old ~2 MiB leaves the samples bunched and fails it;
the chance of 8 genuine 30-bit draws all landing within 1 GiB is ~2^-84.

Not widened: struct task_info still carries a uint32_t eip (cosmetic — ps
shows only your own task, and it is not an ASLR leak), and the legacy
SYS_EXEC raw-base path (no shipped userspace calls it). Both noted rather
than touched, to keep this change to the placement and the truncations it
exposed.

All 21 QEMU suites pass, including the full FS suite (IPC reply buffers at
high bases), smp at -smp 8, and smoke-session. SMP=1 clean; reproducible
build green. Rust unit tests 57 -> 59 landed with the per-task handler
change this builds on.

+++ userspace/sigtest.c
@@ -21,7 +21,7 @@ static void handler(void) {
 }
 
 void _start(void) {
-    if (sys_signal((unsigned)(unsigned long)&handler) != 0) {
+    if (sys_signal((uintptr_t)&handler) != 0) {
         wr("SIGNAL_SELFTEST: FAIL register\n");
         for (;;) { sys_yield(); }
     }
